Fixed the bug of canvas
The bug reported recently is fixed, which would show only part of drawings because of the different pixels and sizes of clients' windows. In this version, no matter how big or small the clients' screens are, the drawings can be shown totally by resizing the drawing. However, new bug is found that the drawing will be scaled and looks different because the different scale of its origin view. Next step, I will try to stick the scale of user's canvas, so that the drawing will stay true scale after transfering.